Address several additional comments, fix some docstrings.
