+++ dingo/gw/data/data_preparation.py
@@ -7,7 +7,8 @@ from dingo.core.dataset import DingoDataset
 from dingo.core.utils.misc import recursive_check_dicts_are_equal
 from dingo.gw.data.data_download import download_raw_data
 from dingo.gw.gwutils import get_window
-from dingo.gw.domains import build_domain_from_model_metadata, UniformFrequencyDomain
+from dingo.gw.domains import UniformFrequencyDomain
+from dingo.gw.domains import build_domain_from_model_metadata
 
 
 def load_raw_data(time_event, settings, event_dataset=None):
@@ -35,7 +36,6 @@ def load_raw_data(time_event, settings, event_dataset=None):
 
     # first try to load the event data from the saved dataset
     if event_dataset is not None:
-
         if isfile(event_dataset):
             dataset = DingoDataset(file_name=event_dataset, data_keys=[event])
             if settings is not None:
@@ -98,7 +98,7 @@ def data_to_domain(raw_data, settings_raw_data, domain, **kwargs):
 
     """
 
-    if type(domain) == UniformFrequencyDomain:
+    if isinstance(domain, UniformFrequencyDomain):
         window = get_window(kwargs["window"])
         data = {"waveform": {}, "asds": {}}
         # convert event strains to frequency domain
