+++ dingo/gw/dataset/generate_dataset.py
@@ -25,6 +25,7 @@ from dingo.gw.waveform_generator import (
 )
 from dingo.core.utils.misc import call_func_strict_output_dim
 
+
 def generate_parameters_and_polarizations(
     waveform_generator: WaveformGenerator,
     prior: BBHPriorDict,
@@ -198,11 +199,14 @@ def generate_dataset(settings: Dict, num_processes: int) -> WaveformDataset:
                 n_test = svd_settings.get("num_validation_samples", 0)
 
                 func = partial(
-                    generate_parameters_and_polarizations, 
-                    waveform_generator, 
+                    generate_parameters_and_polarizations,
+                    waveform_generator,
                     prior,
-                    num_processes=num_processes)
-                parameters, polarizations = call_func_strict_output_dim(func, n_train + n_test)
+                    num_processes=num_processes,
+                )
+                parameters, polarizations = call_func_strict_output_dim(
+                    func, n_train + n_test
+                )
                 svd_dataset_settings = copy.deepcopy(settings)
                 svd_dataset_settings["num_samples"] = len(parameters)
                 del svd_dataset_settings["compression"]["svd"]
@@ -231,11 +235,14 @@ def generate_dataset(settings: Dict, num_processes: int) -> WaveformDataset:
         waveform_generator.transform = Compose(compression_transforms)
 
     func = partial(
-        generate_parameters_and_polarizations, 
-        waveform_generator, 
+        generate_parameters_and_polarizations,
+        waveform_generator,
         prior,
-        num_processes=num_processes)
-    parameters, polarizations = call_func_strict_output_dim(func, settings["num_samples"])
+        num_processes=num_processes,
+    )
+    parameters, polarizations = call_func_strict_output_dim(
+        func, settings["num_samples"]
+    )
     dataset_dict["parameters"] = parameters
     dataset_dict["polarizations"] = polarizations
 
@@ -277,7 +284,6 @@ def parse_args():
 def _generate_dataset_main(
     settings_file: str, out_file: str, num_processes: int
 ) -> None:
-
     if not Path(settings_file).is_file():
         raise FileNotFoundError(f"dataset generation, failed to find {settings_file}")
     if not Path(out_file).parent.is_dir():
