+++ dingo/pipe/main.py
@@ -15,7 +15,7 @@ from bilby_pipe.utils import (
 from .dag_creator import generate_dag
 from .parser import create_parser
 
-from dingo.gw.domains import build_domain_from_model_metadata
+from ..gw.domains.build_domain import build_domain_from_model_metadata
 from dingo.core.posterior_models.build_model import build_model_from_kwargs
 
 logger.name = "dingo_pipe"
