+++ tests/gw/test_noise_dataset.py
@@ -82,7 +82,11 @@ def test_noise_dataset_update_domain(noise_dataset):
     f_max = asd_dataset.domain.f_max
     f_min_new = 20
     f_max_new = 40
-    domain_update = {"f_min": f_min_new, "f_max": f_max_new, "type": "UniformFrequencyDomain"}
+    domain_update = {
+        "f_min": f_min_new,
+        "f_max": f_max_new,
+        "type": "UniformFrequencyDomain",
+    }
     asd_dataset_truncated = ASDDataset(dataset_path, domain_update=domain_update)
     asd_samples_truncated = asd_dataset_truncated.sample_random_asds()
