+++ tests/gw/test_ufd.py
@@ -1,6 +1,7 @@
 import torch
 
-from dingo.gw.domains import UniformFrequencyDomain, TimeDomain, build_domain
+from dingo.gw.domains import UniformFrequencyDomain, TimeDomain
+from dingo.gw.domains import build_domain
 from dingo.gw.gwutils import get_window_factor
 import pytest
 import numpy as np
